refactor: network redesign for the sake of shaving bandwith

+++ Mortz.Core/Net/Messages/CarveMsg.cs
@@ -4,4 +4,4 @@ namespace Mortz.Core.Net.Messages;
 /// owner's client can confirm its predicted carve; 0/-1 for carves nobody
 /// predicted (debug).</summary>
 [NetMessage(NetChannel.RELIABLE, NetDirection.SERVER_TO_CLIENT)]
-public readonly partial record struct CarveMsg(int X, int Y, int Radius, int OwnerId, int SpawnSeq);
+public readonly partial record struct CarveMsg(short X, short Y, byte Radius, int OwnerId, int SpawnSeq);
